Server/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
blinker==1.7.0
boto3==1.29.6
botocore==1.32.6
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
distro==1.8.0
Flask==3.0.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.3
openai==1.3.5
Pillow==10.1.0
pydantic==2.5.2
pydantic_core==2.14.5
python-dateutil==2.8.2
requests==2.31.0
s3transfer==0.7.0
six==1.16.0
sniffio==1.3.0
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==1.26.18
Werkzeug==3.0.1
zipp==3.17.0
